refactor(proj-gen): cleanup phase
