+++ packages/teleport-component-generator/src/index.ts
@@ -3,7 +3,6 @@ import AssemblyLine from './assembly-line'
 import Builder from './builder'
 import Resolver from './resolver'
 
-import { createFile } from '@teleporthq/teleport-shared/lib/utils/project-utils'
 import { camelCaseToDashCase } from '@teleporthq/teleport-shared/lib/utils/string-utils'
 
 import {
@@ -119,6 +118,10 @@ const fileBundler = (fileName: string, codeChunks: Record<string, string>) => {
   const cleanFileName = camelCaseToDashCase(fileName)
 
   return Object.keys(codeChunks).map((fileId) => {
-    return createFile(cleanFileName, fileId, codeChunks[fileId])
+    return {
+      name: cleanFileName,
+      fileType: fileId,
+      content: codeChunks[fileId],
+    }
   })
 }
