+++ packages/teleport-project-generator/src/types.ts
@@ -37,3 +37,20 @@ export interface EntryFileOptions {
   assetsPrefix?: string
   appRootOverride?: string
 }
+
+export interface ComponentGeneratorOutput {
+  files: GeneratedFile[]
+  dependencies: Record<string, string>
+}
+
+export interface PackageJSON {
+  name: string
+  description: string
+  version: string
+  main?: string
+  author?: string
+  license?: string
+  scripts?: Record<string, string>
+  dependencies?: Record<string, string>
+  [key: string]: any
+}
