+++ packages/teleport-project-packer-playground/__tests__/react-basic-variation.json
@@ -32,11 +32,5 @@
       }
     ]
   },
-  "structure": {
-    "componentsPath": [
-      "components",
-      "playground-components"
-    ]
-  },
   "assets": {}
 }
