+++ packages/teleport-project-packer-playground/__tests__/react-next-variation.json
@@ -29,11 +29,5 @@
       }
     ]
   },
-  "structure": {
-    "componentsPath": [
-      "components",
-      "playground-components"
-    ]
-  },
   "assets": {}
 }
