+++ packages/teleport-project-packer-playground/__tests__/vue-basic-variation.json
@@ -32,12 +32,5 @@
       }
     ]
   },
-  "structure": {
-    "componentsPath": [
-      "src",
-      "components",
-      "playground-components"
-    ]
-  },
   "assets": {}
 }
