+++ packages/teleport-project-packer-playground/__tests__/vue-nuxt-variation.json
@@ -26,11 +26,5 @@
       }
     ]
   },
-  "structure": {
-    "componentsPath": [
-      "components",
-      "playground-components"
-    ]
-  },
   "assets": {}
 }
